requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.6.0
django-encrypted-model-fields==0.6.5
django-solo==2.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
Markdown==3.7
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
gunicorn==23.0.0